kernel: Merge Leeco Le1s kernel changes to 3.10.72
Updated using Leeco Le1s kernel (DHXOSOP5801708261S)
Based on ALPS-MP-M0.MP11-V1_VZ6795_LWT_M_KERNEL by @Dinolek

+++ drivers/misc/mediatek/aee/ipanic/ipanic_mtd.c
@@ -222,7 +222,7 @@ static void ipanic_mtd_notify_add(struct mtd_info *mtd)
 	LOGI("Bound to '%s', write size(%d), write size shift(%d)\n",
 	     mtd->name, mtd->writesize, mtd->writesize_shift);
 	ctx->bufsize = ALIGN(PAGE_SIZE, mtd->writesize);
-	ctx->buf = (u64) (int)kmalloc(ctx->bufsize, GFP_KERNEL);
+	ctx->buf = (u64) (unsigned long)kmalloc(ctx->bufsize, GFP_KERNEL);
 	ipanic_mtd_ctx = ctx;
 	return;
